Release 1.0.8
Altered functions involved with consuming workshop resources in WorkshopProductionManager. There is now a V2 of each of them that accepts an argument abCheckOnly for the sake of simply testing if enough resources exist without actually consuming them.
Added GetContainerKeyword and GetContainerForItem functions for querying for the container to use based on the workshop and base item that would be involved with that container.
Added ResourceShortage system.
ResourceShortages are a way for mods to report that a settlement resource is too low. This gives a centralized location for mods to check for additional needs in a settlement beyond the normal rules governing settlements (ie. Food must be higher than population, Defense must be higher than food+water, etc).
This could be used for things like reports, or dynamic systems like auto-assignment, or dynamic resource generation/construction. Ex. Sim Settlements could use this information to prioritize which plots/work objects to build first from a City Plan.
To report a shortage, call RegisterResourceShortage(WorkshopScript akWorkshopRef, ActorValue aResourceAV, Float afTargetWorkshopValue) on WorkshopResourceManager.
Note that afTargetWorkshopValue is the total value it should have for that shortage to be considered resolved. Ex. If your mod logic requires settlement food be at 40 to no longer be considered short, you would send WorkshopResourceManager.RegisterResourceShortage(someWorkshopRef, FoodAV, 40.0).
Only the highest shortage for each AV is stored. For example, if you report a shortage of 2 food, and another mod reports a shortage of 4 food, the shortage of 4 will take priority. If a higher shortage already exists, RegisterResourceShortage will return false, otherwise it will return true. (Note it will also return false if akWorkshopRef or aResourceAV are None)
ResourceShortages will be tested after new workshop objects are built, on demand when WorkshopResourceManager.UpdateResourceShortages(WorkshopScript akWorkshopRef) is called, and daily on a timer.
The test checks for two things: Is the shortage resolved, has the shortage expired.
A shortage is resolved if the workshop’s AV matches or exceeds the amount reported as needed.
A shortage expires after 2 in-game days. This is to ensure that a shortage doesn’t become permanent if circumstances change and the mod that registered doesn’t correctly refresh the shortage.
Registering the shortage again will refresh it. Mods should plan on re-registering shortages periodically, or reacting to the ResourceShortageExpired event in order to ensure their shortages remain relevant.
Only the largest shortage for each AV is stored.
Once a shortage is resolved a custom event (ResourceShortageResolved) will be fired with the workshop, ActorValue, and current value of that AV. Mods can register for this Event to react to the resolution.
Once a shortage expires, a custom event (ResourceShortageExpired) will be fired with the workshop, ActorValue, and current value of that AV. Mods can register for this Event to react to the expiration (perhaps by re-registering the shortage if it still exists according to the mod’s gameplay logic).
ScrapObject thread will now move items deep below the world before enabling them to test for AVs to hide this from the player.
Added a MessageManager quest which allows the player to control circumstances when messages show up. Many workshop messages have been routed through this system, and any mod may make use of it as well to display their own messages.
If a message is sent and the player’s settings are set to suppress that message, for example - they can arrange so messages don’t come up during combat, they will instead be queued.
Depending on the player’s settings, Queued messages will either be displayed shortly after circumstances allow them to, or when triggered to display via a holotape control.
All raid messages from the base game are routed through this system.
Filled the missing script properties in the APIQuest.