CollectPoweredItems bug fix
If the number of wires or powered objects found exceeds the array capacity, the player is presented a warning message. The issue is the loop keeps iterating the found linked refs and the player is presented the warning message each iteration leaving them to think the game is hanging.

I added 2 bools to use as break conditions so when the arrays are full, it will no longer check that object type. The other option would be to break the function in half. One checking for wires the other checking for powerable objects. I figure the former is a better option.

Also, it was reported in the forum here:
https://simsettlements.com/site/index.php?threads/faux-power-doesnt-power-wireless-lights.23113
It seems you had one user exceed the powerable object array length. I'm really curious how you could have more than 1024 objects in a settlement without the power of Todd running rampant...