StoreArguments in ThreadRunner
StoreArguments function is way more complicated than it needs to be.
Var[] akArgs has already been created. The array does not need to be init, it can just be set to akArgs.
This also eliminates the need for the complicated array length adjustment code ant the end of the function as currently written.